crawler/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-cloud-firestore>=2.11.0
lxml>=4.9.0
feedparser>=6.0.0
google-auth>=2.0.0